+++ src/MR.EntityFrameworkCore.KeysetPagination/KeysetPaginationExtensions.cs
@@ -1,7 +1,6 @@
 ﻿using System.Linq.Expressions;
 using System.Reflection;
 using Microsoft.EntityFrameworkCore;
-using Microsoft.EntityFrameworkCore.Metadata.Internal;
 
 namespace MR.EntityFrameworkCore.KeysetPagination;